backend/requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
umap-learn
numpy
pandas
